requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==5.2
django-ninja==1.4.1
djantic==0.7.0
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.10
pydantic==2.11.4
pydantic_core==2.33.2
sqlparse==0.5.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
